Day/night cycle works for textures and world light

+++ Assignment/Light.java
@@ -176,13 +176,12 @@ public class Light {
         gl.glDeleteBuffers(1, elementBufferId, 0);
     }
 
-    public void setPower(int lightNum, int powerLevel) {
+    public void setPower(int lightNum, float powerLevel) {
         material.setDiffusePoint(lightNum, originalDiffuse.get(lightNum).x*powerLevel, originalDiffuse.get(lightNum).y*powerLevel, originalDiffuse.get(lightNum).z*powerLevel);
         material.setSpecularPoint(lightNum, originalSpecular.get(lightNum).x*powerLevel, originalSpecular.get(lightNum).y*powerLevel, originalSpecular.get(lightNum).z*powerLevel);
         if (lightNum != 3) {
             lampColor = powerLevel + 0.1f;
         }
-
     }
 
     public String toString() {
